requirements.txt
flask==2.2.5
flask-bootstrap==3.3.7.1
cassandra-driver==3.25.0
apscheduler==3.10.1
